requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-model-utils==4.0.0
django-tinymce4-lite==1.8.0
gast==0.2.2
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
jsmin==2.2.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
mysqlclient==1.4.6
numpy==1.18.2
oauthlib==3.1.0
opencv-python-headless==4.2.0.34
opt-einsum==3.2.0
Pillow==7.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.12.1
